feat(entityheader): supports v3
